open-telemetry/arize/pyproject.toml
[project]
name = "pipecat-open-inference-tracing"
version = "0.1.0"
description = "A Pipecat example using OpenInference tracing"
requires-python = ">=3.11"
dependencies = [
    "arize-otel>=0.11.0",
    "boto3>=1.34.0",
    "openinference-instrumentation>=0.1.42",
    "openinference-instrumentation-pipecat>=0.1.0",
    "pipecat-ai[cartesia,daily,deepgram,google,runner,silero,webrtc,websocket]>=0.0.100",
    "pipecatcloud>=0.4.4",
]

[dependency-groups]
dev = [
    "ruff~=0.12.1",
]

phoenix = [
    "arize-phoenix>=12.30.0",
]

[tool.ruff]
line-length = 100
[tool.ruff.lint]
select = ["I"]
